Move case values tests into a single file.
